backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
